+++ R/mvt.R
@@ -1,0 +1,9 @@
+## mapbox vector tiles format
+
+#' Mapbox attribution
+#'
+#' @export
+mapbox_attribution <- function() { # nocov start
+  c("&copy; <a href='https://www.mapbox.com/about/maps/'>Mapbox</a>, ",
+    "&copy; <a href='http://www.openstreetmap.org/copyright'>OpenStreetMap</a>")
+} # nocov end